Agent6/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
huggingface_hub>=0.23.0
python-dotenv>=1.0.0
